MEV-1868 update FuluAdjustableSubmitBlockRequest
