+++ bidadjustment/types.go
@@ -52,46 +52,6 @@ func (v *VersionedAdjustmentData) IsEmpty() bool {
 	}
 }
 
-func (v *VersionedAdjustmentData) MarshalSSZ() ([]byte, error) {
-	switch v.Version {
-	case AdjustmentDataVersion3:
-		return v.V3.MarshalSSZ()
-	case AdjustmentDataVersion2:
-		return v.V2.MarshalSSZ()
-	case AdjustmentDataVersion1:
-		return v.V1.MarshalSSZ()
-	default:
-		return nil, errors.Wrap(ErrInvalidVersion, fmt.Sprintf("%d is not supported", v.Version))
-	}
-}
-
-func (v *VersionedAdjustmentData) UnmarshalSSZ(input []byte) error {
-	var err error
-
-	dataV3 := new(AdjustmentDataV3)
-	if err = dataV3.UnmarshalSSZ(input); err == nil {
-		v.Version = AdjustmentDataVersion3
-		v.V3 = dataV3
-		return nil
-	}
-
-	dataV2 := new(AdjustmentDataV2)
-	if err = dataV2.UnmarshalSSZ(input); err == nil {
-		v.Version = AdjustmentDataVersion2
-		v.V2 = dataV2
-		return nil
-	}
-
-	dataV1 := new(AdjustmentData)
-	if err = dataV1.UnmarshalSSZ(input); err == nil {
-		v.Version = AdjustmentDataVersion1
-		v.V1 = dataV1
-		return nil
-	}
-
-	return errors.Wrap(err, "failed to unmarshal VersionedAdjustmentData SSZ")
-}
-
 func (v *VersionedAdjustmentData) MarshalJSON() ([]byte, error) {
 	switch v.Version { //nolint:exhaustive
 	case AdjustmentDataVersion3:
@@ -573,9 +533,10 @@ type FuluAdjustableSubmitBlockRequest struct {
 	BlobsBundle       *fulu.BlobsBundle
 	ExecutionRequests *electra.ExecutionRequests
 	Signature         phase0.BLSSignature `ssz-size:"96"`
-	AdjustmentData    *AdjustmentData
+	AdjustmentData    *VersionedAdjustmentData
 }
 
+// TODO: hopefully won't need versioning functions after this PR is done
 func AdjustmentDataV1ToVersioned(adjustmentData *AdjustmentData) *VersionedAdjustmentData {
 	if adjustmentData == nil {
 		return nil
